Remove collections npm package
SortedMap is a tree, and for some reason it exhibited pathological behaviour on EVERY RUN apparently, becoming a list instead. And iterating a list of 200000 items using a recursive function... that's a bit of a no-no in JS. So apparently proper standard collections isn't a solved thing in JS, boo!